moved indexing logic into converter specific utils folder
