+++ packages/data-sql/src/query-converter/fields/function.ts
@@ -1,6 +1,6 @@
 import type { AbstractQueryFunction } from '@directus/data';
 import type { AbstractSqlQuerySelectFnNode, ParameterTypes } from '../../types/index.js';
-import type { IndexGenerators } from '../../utils/create-index-generators.js';
+import type { IndexGenerators } from '../utils/create-index-generators.js';
 import { convertFn } from '../common/function.js';
 
 /**
