+++ packages/data-sql/src/query-converter/modifiers/filter/conditions/conditions.test.ts
@@ -1,7 +1,7 @@
 import type { AbstractQueryConditionNode } from '@directus/data';
 import { randomAlpha, randomIdentifier, randomInteger } from '@directus/random';
 import { expect, test } from 'vitest';
-import { createIndexGenerators } from '../../../../utils/create-index-generators.js';
+import { createIndexGenerators } from '../../../utils/create-index-generators.js';
 import type { FilterResult } from '../utils.js';
 import { convertCondition } from './conditions.js';
