+++ packages/data-sql/src/query-converter/modifiers/filter/conditions/field.test.ts
@@ -1,7 +1,7 @@
 import type { ConditionFieldNode } from '@directus/data';
 import { randomIdentifier, randomInteger } from '@directus/random';
 import { expect, test } from 'vitest';
-import { createIndexGenerators } from '../../../../utils/create-index-generators.js';
+import { createIndexGenerators } from '../../../utils/create-index-generators.js';
 import type { FilterResult } from '../utils.js';
 import { convertFieldCondition } from './field.js';
