+++ packages/data-sql/src/query-converter/modifiers/filter/conditions/geo.test.ts
@@ -2,7 +2,7 @@ import type { ConditionGeoIntersectsBBoxNode } from '@directus/data';
 import { randomIdentifier, randomInteger } from '@directus/random';
 import { expect, test } from 'vitest';
 import type { GeoJSONGeometry } from 'wellknown';
-import { createIndexGenerators } from '../../../../utils/create-index-generators.js';
+import { createIndexGenerators } from '../../../utils/create-index-generators.js';
 import type { FilterResult } from '../utils.js';
 import { convertGeoCondition } from './geo.js';
