+++ packages/data-sql/src/query-converter/modifiers/filter/conditions/number.ts
@@ -1,5 +1,5 @@
 import type { ConditionNumberNode } from '@directus/data';
-import type { IndexGenerators } from '../../../../utils/create-index-generators.js';
+import type { IndexGenerators } from '../../../utils/create-index-generators.js';
 import { convertTarget } from '../../target.js';
 import type { FilterResult } from '../utils.js';
