+++ packages/data-sql/src/query-converter/modifiers/filter/conditions/string.test.ts
@@ -1,7 +1,7 @@
 import type { ConditionStringNode } from '@directus/data';
 import { randomAlpha, randomIdentifier, randomInteger } from '@directus/random';
 import { expect, test } from 'vitest';
-import { createIndexGenerators } from '../../../../utils/create-index-generators.js';
+import { createIndexGenerators } from '../../../utils/create-index-generators.js';
 import type { FilterResult } from '../utils.js';
 import { convertStringNode } from './string.js';
