+++ packages/data-sql/src/utils/index.ts
@@ -1,6 +1,4 @@
-export * from './create-index-generators.js';
 export * from './get-mapped-queries-stream.js';
 export * from './map-result.js';
-export * from './number-generator.js';
 export * from './numeric-operator-conversion.js';
 export * from './stream-consumer.js';
